stats and bounty now int + improved global stats gathering

+++ src/main/we/rashchenko/patronum/database/stats/GlobalStats.kt
@@ -3,8 +3,8 @@ package we.rashchenko.patronum.database.stats
 import kotlin.math.abs
 
 class GlobalStats {
-    var sortedReputations: Array<Float> = arrayOf()
-    fun getRateOnLeaderBoard(reputation: Float): Float{
+    var sortedReputations: Array<Int> = arrayOf()
+    fun getRateOnLeaderBoard(reputation: Int): Float{
         val place = sortedReputations.binarySearch(reputation)
         return abs(place).toFloat() / sortedReputations.size
     }
